Fixed makePlatformPart method
