+++ src/chrome/latest.ts
@@ -2,10 +2,10 @@ import { InstallerFactory } from "./installerFactory";
 import * as hc from "@actions/http-client";
 import * as tc from "@actions/tool-cache";
 import { SnapshotInstaller } from "./snapshot";
-import { makePlatformPart } from "../utils";
 import { getPlatform } from "../platform";
 import { InstallResult } from "../installer";
 import { DownloadResult } from "../downloadUrl";
+import { makePlatformPart } from "./utils";
 
 export class LatestInstaller implements InstallerFactory {
   private readonly http = new hc.HttpClient("abhi1693/setup-browser");
