+++ src/chrome/utils.ts
@@ -1,4 +1,4 @@
-import { OS, Platform } from "../platform";
+import { Arch, OS, Platform } from "../platform";
 
 export const makeBasename = ({ os }: Platform): string => {
   switch (os) {
@@ -9,4 +9,19 @@ export const makeBasename = ({ os }: Platform): string => {
     case OS.WINDOWS:
       return "chrome-win.zip";
   }
+};
+
+export const makePlatformPart = ({ os, arch }: Platform): string => {
+  if (os === OS.DARWIN && arch === Arch.AMD64) {
+    return "Mac";
+  } else if (os === OS.LINUX && arch === Arch.I686) {
+    return "Linux";
+  } else if (os === OS.LINUX && arch === Arch.AMD64) {
+    return "Linux_x64";
+  } else if (os === OS.WINDOWS && arch === Arch.I686) {
+    return "Win";
+  } else if (os === OS.WINDOWS && arch === Arch.AMD64) {
+    return "Win_x64";
+  }
+  throw new Error(`Unsupported platform "${os}" "${arch}"`);
 };
